placed add, edit, and delete book test in owner library test
