+++ app/src/androidTest/java/com/example/glassesgang/OwnerLibraryTest.java
@@ -1,6 +1,5 @@
 package com.example.glassesgang;
 
-import android.view.View;
 import android.widget.EditText;
 
 import androidx.test.platform.app.InstrumentationRegistry;
@@ -20,13 +19,14 @@ import org.junit.Test;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-public class EditDeleteBookTest {
+public class OwnerLibraryTest {
     private static FirebaseAuth mAuth;
     private static DatabaseManager databaseManager;
     private String email = "mockuser@gmail.com";
     private String password = "12345678";
     private Solo solo;
     private int timeout = 3000;
+
     @Rule
     public ActivityTestRule<MainActivity> rule =
             new ActivityTestRule<>(MainActivity.class, true, true);
@@ -40,7 +40,9 @@ public class EditDeleteBookTest {
         // Initialize Firebase Auth
         mAuth = FirebaseAuth.getInstance();
         databaseManager = new DatabaseManager();
-        mAuth.signOut();
+        if (mAuth.getCurrentUser() != null) {
+            mAuth.signOut();
+        }
     }
 
     /**
@@ -75,15 +77,19 @@ public class EditDeleteBookTest {
         databaseManager.clearOwnerCatalogue(email);
     }
 
+    /**
+     * manually signs in the mock user
+     */
     public void signInMockUser() {
         solo.enterText((EditText) solo.getView(R.id.email), email);
         solo.enterText((EditText) solo.getView(R.id.password), password);
         solo.clickOnButton("SIGN IN");
-        solo.waitForActivity(OwnerHomeActivity.class, timeout);
+        solo.waitForActivity(OwnerHomeActivity.class);
+        solo.clickOnText("Books");
     }
 
     /**
-     * adding a test Book
+     * manually adds a dummy book
      * @param title title of the book to add
      * @param author author of the book to add
      * @param isbn isbn of the book to add
@@ -96,7 +102,139 @@ public class EditDeleteBookTest {
         solo.enterText((EditText) solo.getView(R.id.isbn_bar), isbn);
         solo.clickOnButton("SAVE");
 
-        solo.waitForText(title, 1, timeout);
+        solo.waitForText(title, 1, timeout);  // waiting for the book to be added on the list view
+    }
+
+    /**
+     * Checks if pressing the plus button directs user to add book activity
+     */
+    @Test
+    public void checkAddButton() {
+        solo.clickOnView(solo.getView(R.id.add_button));
+        solo.assertCurrentActivity("Wrong activity after pressing add button", AddBookActivity.class);
+    }
+
+    /**
+     * Checks if added book goes into the list view
+     */
+    @Test
+    public void checkListViewAfterAdding() {
+        String title = "Title";
+        String author = "AuthorFirst AuthorLast";
+        String isbn = "1234567890123";
+
+        addBook(title, author, isbn);
+
+        // check if the book is added in the list view
+        solo.assertCurrentActivity("Wrong activity", OwnerHomeActivity.class);
+        assertTrue(solo.waitForText(title, 1, timeout));
+        assertTrue(solo.waitForText(author, 1, timeout));
+        assertTrue(solo.waitForText(isbn, 1, timeout));
+    }
+
+    /**
+     * Checks if pressing back button returns to owner home activity
+     * and that it does not add any book
+     */
+    @Test
+    public void testBackButton() {
+        String title = "Title";
+        String author = "AuthorFirst AuthorLast";
+        String isbn = "1234567890123";
+
+        // verify that the book is not in the list view in the first place
+        assertFalse(solo.waitForText(title, 1, 2000));
+        assertFalse(solo.waitForText(author, 1, 2000));
+        assertFalse(solo.waitForText(isbn, 1, 2000));
+
+        solo.clickOnView(solo.getView(R.id.add_button));
+        solo.assertCurrentActivity("Wrong activity after pressing add button", AddBookActivity.class);
+
+        // enter book description, but don't save
+        solo.enterText((EditText) solo.getView(R.id.book_title_bar), title);
+        solo.enterText((EditText) solo.getView(R.id.author_name_bar), author);
+        solo.enterText((EditText) solo.getView(R.id.isbn_bar), isbn);
+        solo.clickOnButton("BACK");
+
+        // make sure the book is not added in the list view
+        solo.assertCurrentActivity("Wrong activity after pressing BACK button", OwnerHomeActivity.class);
+        assertFalse(solo.waitForText(title, 1, 2000));
+        assertFalse(solo.waitForText(author, 1, 2000));
+        assertFalse(solo.waitForText(isbn, 1, 2000));
+    }
+
+    /**
+     * Check whether the app prevents the user from adding books
+     * when they don't provide any of the author, isbn or title
+     */
+    @Test
+    public void checkEmptyFieldVerification() {
+        String title = "Title";
+        String author = "AuthorFirst AuthorLast";
+        String isbn = "1234567890123";
+
+        solo.clickOnView(solo.getView(R.id.add_button));
+        solo.assertCurrentActivity("Wrong activity after pressing add button", AddBookActivity.class);
+
+        EditText titleField = (EditText) solo.getView(R.id.book_title_bar);
+        EditText authorField = (EditText) solo.getView(R.id.author_name_bar);
+        EditText isbnField = (EditText) solo.getView(R.id.isbn_bar);
+
+
+        // leaving isbn field blank
+        solo.enterText(titleField, title);
+        solo.enterText(authorField, author);
+        solo.clickOnButton("SAVE");
+        solo.assertCurrentActivity("Wrong activity", AddBookActivity.class);
+
+        // leaving author field blank
+        solo.clearEditText(authorField);
+        solo.enterText(isbnField, isbn);
+        solo.clickOnButton("SAVE");
+        solo.assertCurrentActivity("Wrong activity", AddBookActivity.class);
+
+        // leaving the title field blank
+        solo.clearEditText(titleField);
+        solo.enterText(authorField, author);
+        solo.clickOnButton("SAVE");
+        solo.assertCurrentActivity("Wrong activity", AddBookActivity.class);
+
+        // leaving all fields blank
+        solo.clearEditText(authorField);
+        solo.clearEditText(isbnField);
+        solo.clickOnButton("SAVE");
+        solo.assertCurrentActivity("Wrong activity", AddBookActivity.class);
+    }
+
+    /**
+     * Checks whether the app prevents user to add books with invalid isbn
+     */
+    @Test
+    public void testISBNFieldVerification() {
+        String title = "Title";
+        String author = "AuthorFirst AuthorLast";
+        String longISBN = "12345678901231234567890123";
+        String shortISBN = "123";
+
+        solo.clickOnView(solo.getView(R.id.add_button));
+        solo.assertCurrentActivity("Wrong activity after pressing add button", AddBookActivity.class);
+
+        EditText isbnField = (EditText) solo.getView(R.id.isbn_bar);
+
+        // enter book description
+        solo.enterText((EditText) solo.getView(R.id.book_title_bar), title);
+        solo.enterText((EditText) solo.getView(R.id.author_name_bar), author);
+
+        // enter invalid long isbn
+        solo.enterText(isbnField, longISBN);
+        solo.clickOnButton("SAVE");
+        solo.assertCurrentActivity("Wrong activity", AddBookActivity.class);
+
+        // enter invalid short isbn
+        solo.clearEditText(isbnField);
+        solo.enterText(isbnField, shortISBN);
+        solo.clickOnButton("SAVE");
+        solo.assertCurrentActivity("Wrong activity", AddBookActivity.class);
     }
 
     /**
@@ -107,7 +245,10 @@ public class EditDeleteBookTest {
         String title = "Original Title";
         String author = "Original Author";
         String isbn = "1234567890123";
+
         addBook(title, author, isbn);
+
+        // click on the added book
         solo.clickOnText(title);
         solo.assertCurrentActivity("Wrong Activity after clicking book", OwnerBookProfileActivity.class);
 
@@ -115,6 +256,28 @@ public class EditDeleteBookTest {
         assertTrue(solo.waitForText(title, 1, timeout));
         assertTrue(solo.waitForText(author, 1, timeout));
         assertTrue(solo.waitForText(isbn, 1, timeout));
+    }
+
+    @Test
+    public void testEditFieldHasBookDesc() {
+        String origTitle = "Original Title";
+        String origAuthor = "Original Author";
+        String origISBN = "1234567890123";
+
+        // add a book and open up its profile
+        addBook(origTitle, origAuthor, origISBN);
+        solo.clickOnText(origTitle);
+        solo.assertCurrentActivity("Wrong Activity after clicking book", OwnerBookProfileActivity.class);
+
+        // click edit button
+        solo.clickOnButton("EDIT");
+        solo.assertCurrentActivity("Wrong Activity after clicking book", OwnerBookProfileActivity.class);
+
+        // check if the edit text fields contain the book info
+        solo.assertCurrentActivity("Wrong Activity", EditBookActivity.class);
+        assertTrue(solo.waitForText(origTitle, 1, timeout));
+        assertTrue(solo.waitForText(origAuthor, 1, timeout));
+        assertTrue(solo.waitForText(origISBN, 1, timeout));
     }
 
     /**
@@ -136,17 +299,16 @@ public class EditDeleteBookTest {
 
         // click edit button
         solo.clickOnButton("EDIT");
-        solo.assertCurrentActivity("Wrong Activity after clicking book", OwnerBookProfileActivity.class);
+        solo.assertCurrentActivity("Wrong Activity after clicking book", EditBookActivity.class);
+        solo.waitForText(origTitle, 1, 5000); // waiting for the orig description to show before clearing text fields
 
         // edit the book and save changes
         EditText titleField = (EditText) solo.getView(R.id.book_title_bar);
         EditText authorField = (EditText) solo.getView(R.id.author_name_bar);
         EditText isbnField = (EditText) solo.getView(R.id.isbn_bar);
-        solo.waitForText(origTitle, 1, timeout);
         solo.clearEditText(titleField);
         solo.clearEditText(authorField);
         solo.clearEditText(isbnField);
-        solo.waitForText("", 3, timeout);
         solo.enterText(titleField, editedTitle);
         solo.enterText(authorField, editedAuthor);
         solo.enterText(isbnField, editedISBN);
@@ -178,23 +340,19 @@ public class EditDeleteBookTest {
 
         // click edit button
         solo.clickOnButton("EDIT");
-        solo.assertCurrentActivity("Wrong Activity after clicking book", OwnerBookProfileActivity.class);
+        solo.assertCurrentActivity("Wrong Activity after clicking book", EditBookActivity.class);
+        solo.waitForText(origTitle, 1, 5000); // waiting for the orig description to show before clearing text fields
 
         // edit the book and save changes
         EditText titleField = (EditText) solo.getView(R.id.book_title_bar);
         EditText authorField = (EditText) solo.getView(R.id.author_name_bar);
         EditText isbnField = (EditText) solo.getView(R.id.isbn_bar);
-        solo.waitForText(origTitle, 1, timeout);
         solo.clearEditText(titleField);
         solo.clearEditText(authorField);
         solo.clearEditText(isbnField);
-        solo.waitForText("", 3, timeout);
         solo.enterText(titleField, editedTitle);
         solo.enterText(authorField, editedAuthor);
         solo.enterText(isbnField, editedISBN);
-        solo.waitForText(editedTitle, 1, timeout);
-        solo.waitForText(editedAuthor, 1, timeout);
-        solo.waitForText(editedISBN, 1, timeout);
         solo.clickOnButton("SAVE");
 
         // return to the list view
@@ -209,7 +367,7 @@ public class EditDeleteBookTest {
     }
 
     /**
-     * Checks if deleteing a book is revmoed from the list view
+     * Checks if deleting a book is removes it from the list view
      */
     @Test
     public void testBookRemovedFromListView() {
@@ -218,7 +376,6 @@ public class EditDeleteBookTest {
         String author = "Author";
         String isbn = "1234567890123";
         addBook(title, author, isbn);
-        solo.waitForText(title, 1, timeout);
 
         // delete the added book
         solo.clickOnText(title);
@@ -227,7 +384,7 @@ public class EditDeleteBookTest {
         solo.clickOnText("Confirm");
 
         // check if the book is removed from the list view
-        assertTrue(solo.waitForActivity(OwnerBookProfileActivity.class, timeout));
+        solo.assertCurrentActivity("Wrong Activity", OwnerHomeActivity.class);
         assertFalse(solo.waitForText(title, 1, timeout));
         assertFalse(solo.waitForText(author, 1, timeout));
         assertFalse(solo.waitForText(isbn, 1, timeout));
@@ -254,10 +411,13 @@ public class EditDeleteBookTest {
         // check that book is not deleted
         solo.assertCurrentActivity("Wrong Activity", OwnerBookProfileActivity.class);
         solo.clickOnActionBarHomeButton();
+        solo.assertCurrentActivity("Wrong Activity", OwnerHomeActivity.class);
         assertTrue(solo.waitForText(title, 1, timeout));
         assertTrue(solo.waitForText(author, 1, timeout));
         assertTrue(solo.waitForText(isbn, 1, timeout));
     }
+
+
 
 
 }
